+++ buraq/models.py
@@ -19,6 +19,9 @@ from buraq.orm.base import CheckConstraint, Index, Model, UniqueConstraint
 from buraq.utils.choices import IntegerChoices, TextChoices
 from buraq.orm.fields import (
     CASCADE,
+    DB_CASCADE,
+    DB_SET_DEFAULT,
+    DB_SET_NULL,
     DO_NOTHING,
     PROTECT,
     RESTRICT,